Rename some files to keep backward compatibility
